Started Rewriting in the style of having many functions
started re-writing the code in functiony language that I head was much more propper/better hoping it would make the effects implementation more obvious, didn't get to that part but finished the timestamps file creation
New code is in v.03 folder. I also switched from notepad++ to vscode for this. Coding is fun!