tldr: started over again but much faster/better because of what i've learned
yeah I got streesed out on the project,
couldn't wrap my head around adding a config for some reason + complexity
I think part of that was self imposed pressure to do well or smthn partially for resume & stuff.
anyway I had trouble working on it including today
until today I quit/decided to procrastinated it
but later the same day I just wanted something quick and dirty
so I could start recording yt videos and maybe I'd keep it private
I tried working it as fast as possible

but of course what ended up happening is I did a better job than on the other project and I remembered to organize it properly and make it a bit simpler.
I copied some code from old projects and also the main architecture design from the old version so that all went much faster I felt like it didnt go to waste. Also comparing the first version I think this is better that was me learning
As for the config file, I do not have one but I have static dicts so all it is is a matter of implementing a function that reads a file and returns those dicts,
Idk why it was hard for me to implement that in the older version, I'm sure its similar.
So then even later this same day I went back to work on it and its much easier.
Side note: also another responsibility that was on my mind was made concrete / planned for so that could effect that stuff too. I was getting tired of responsibility and also I partially let go doing the code as a chore. I just want to make a cool yt video and I am waiting on this. Hopefully I keep that mindset I guess of not too much pressure.
but now I am fully tired & it is 2 am. goodnight :)
Will this commit in 1 go this time vscode?